pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "devops-incident-response"
version = "1.0.0"
description = "DevOps Incident Response OpenEnv"
authors = [
    { name = "devops-incident-env" }
]
requires-python = ">=3.11"
dependencies = [
    "openenv-core>=0.2.3",
    "pydantic>=2.0,<3.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0"
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
py-modules = ["__init__", "client", "models", "env", "inference", "api", "check_all", "check_grader", "validate"]
packages = ["server", "tasks", "graders"]
